+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Converted lookup_accession_numbers script into C++ program w/ memory mapping
 - Clarified in manual that one or more libraries allowed for custom DBs
 - Silenced progress messages in C++ programs for non-TTY stderr
+- Taxonomy downloads switched to rsync from wget (ftp)
 
 ### Fixed
 - Allow d/l of protozoa library w/ kraken2-build script